Use userspace architecture string, not kernel uname
We shouldn't assume a kernel architecture maps to a given userspace
environment. We could be running in an armhf chroot on an aarch64 kernel,
for example. This probably isn't a practical issue for the KVM tests
since we probably wouldn't be certifying such setups, but it does
let us be more explicit about arch detection (vs. relying on regex
matching).