providers:checkbox: Remove nonexisting dependency from some after-suspend jobs
These jobs were renamed and their names are variable, so we can't
express them in another job's depends: as that doesn't support
wildcards.

The easiest workaround at this point is to remove them from the
dependencies. THe consequence is that the affected jobs (after suspend)
will run even if their before-suspend counterpart failed. This doesn't
sound too critical though, as if they fail before suspend but pass after
suspend it may warrant a closer look.