provider:checkbox: fix bad error case in camera_test
This patch changes CameraTest._get_supported_resolutions() not to
return early from the per-format loop when an IOError occurs while
manipulating the device. This allows the 'resolutions' element
to be assigned which ensures that every dictionary in supported_formats
has that element set.

This bug can happen, for example, when the current user has no permission
to manipulate the camera device and gets an IOError with EPERM inside.

Fixes: https://bugs.launchpad.net/plainbox-provider-checkbox/+bug/1235426

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>