providers:checkbox: Fix template and exception handling in xrandr_cycle
The script now looks for its template file in several locations, is more
strict with exception handling, and provides better error messages if
things don't work as expected.

Due to bin/lib split in plainbox providers, it needs a bit of
configuration to be run manually (i.e. from outside checkbox or
plainbox), but the script will provide guidance on the environment
variable to set for this to work.