providers:checkbox: use python3 for auto-running pm_test
This patch corrects auto-start machinery to use python3, not python2 to
run the pm_test script.

The pm_test script sets up automatic execution of itself to cyclic
testing. The way it handles that is by creating an autostart file and by
letting the current user run python as root. That was all correct when
the script was originally written in python2.x, now with python3 it
causes the script to fail.

Fixes: https://bugs.launchpad.net/plainbox-provider-checkbox/+bug/1413222
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>