providers:checkbox: flag jobs using pm_test with 'noreturn'
This patch attaches 'noreturn' flag to all tests using pm_test helper, as this
helper runs in a non-blocking fashion. In case of pm_test those commands are
reboot and poweroff, all of which killing plainbox along the way. If 'noreturn'
flag is not used, this killing might come while plainbox continues to operate
which may result to messed-up session state.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1403832
Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>