provivder:checkbox: Add system logs to stress tests logs
As explained in lp:1584607, more information is needed to investigate some
issues when running stress tests (or tests using fwts_test or pm_test scripts).

The content of dmesg and /var/log/syslog is added to the logs generated by
these scripts.

Since the resulting log files become really big, the attachment jobs are
modified to compress (using xz) and encode the result in base64 because of a
limitation in our system (all attachments must be strings, not binary).