providers:checkbox: Adapt graphics jobs to take amdgpu-pro drivers into account
The amdgpu-pro drivers are based on the open source amdgpu drivers (so udev
shows the same driver name for both amdgpu and amdgpu-pro installs), but use a
dedicated script (amdgpu-pro-px) to switch between integrated and discrete GPUs.

We adapt the existing jobs and scripts to take this into account, based on
Sylvain's modification to graphics_card resource job that allows to show
'amdgpu-pro' as driver name when the amdgpu-pro metapackage is present on the
system.

Fixes: https://bugs.launchpad.net/plainbox-provider-certification-client/+bug/1622477