Disable fwts' maxfreq test on arm64
LP: #1698164

The maxfreq test in fwts is guarded by an "#ifdef FWTS_ARCH_INTEL", making
it only available on x86. I checked the test to see if that still makes
sense, and it does. The test relies on frequency information being available
in /proc/cpuinfo - and that is architecture specific (ARM does not do that).