Add manual test to check decryption error when TPM data modified
There is no easy way to automate a test to check that device data
partition cannot be decrypted if TPM2 data is modified.

The easiest way is to manually flash the BIOS after installing the OEM
image. By doing so, the TPM data, including data used for data partition
decryption, is modified, making it impossible to decrypt the partition
and boot the device.