unit: Add test job to validate power button event filtering (LP: #1877084)
Current WinOS requires ODM vendor to report press and release events
with one power button press, the power button release can be filtered
if 5-button-array is supported.

There is a timing issue that two consecutive and close power button event
might trigger the system goes to shutdown directly wihtout endSessionDialog
pop-up, which is not the expected behaviour.

Create a job for knowing if the system will hit the issue.