delay zapperized insertion/removal for 5s
Previously the command to connect or disconnect the USB device was sent
almost at the same time that removable_watcher started waiting for that
change to take place. This meant that there was a big chance for that
action to happen _before_ the test could observe the change.

This patch introduces a delay (with a background thread) that calls
Zapper after 5 seconds from starting the observation. The whole test is
programmed to wait for 20s, so the delay doesn't change the duration of
the test. It's also not some fixing-race-conditions-with-random-sleeps
kind of fix.